requirements.txt
anyio==3.6.2
APScheduler==3.9.1
autopep8==1.7.0
beautifulsoup4==4.11.1
cachetools==5.0.0
certifi==2022.9.24
charset-normalizer==2.1.1
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
idna==3.4
lxml==4.9.1
numpy==1.23.4
pycodestyle==2.9.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-telegram-bot==20.0a0
python-telegram-bot-raw==13.14
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
toml==0.10.2
tornado==6.1
tzdata==2022.5
tzlocal==4.2
urllib3==1.26.12
